Adds node workspace to the leader
When testing kubernetes, the leader magically worked due to it hosting the apiserver bound to all interfaces on the host, listening on port 8080. This is not the case moving forward with how we've split apart leader and follower nodes.

This necessitates the creation, and usage of the "node workspace" - a la ~/.kube/config.